requirements.txt
Django==1.8.4
django-modeltranslation==0.10.1
djangorestframework==3.2.4
django-ckeditor==5.0.2
